Re-organize dictum directory
