+++ frontend/src/dictum/list/DictumRow.js
