+++ frontend/src/dictum/list/DictumTable.js
