+++ frontend/src/process/list/FilterableProcessTable.js
@@ -1,6 +1,6 @@
 import React, {Component} from "react";
-import ProcessTable from "./list/ProcessTable";
-import {BASE_PATH, restDelete, restGet} from "../commons/RestUtils";
+import ProcessTable from "./ProcessTable";
+import {BASE_PATH, restDelete, restGet} from "../../commons/RestUtils";
 import {Link} from "react-router-dom";
 
 export default class FilterableProcessTable extends Component {
